+++ .eslintignore
@@ -1,0 +1,6 @@
+
+**/node_modules/*
+**/vendor/*
+**/*.min.js
+**/coverage/*
+**/build/*
